wallet and refreshtoken module added

+++ src/admin/admin.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpStatus, Injectable } from '@nestjs/common';
 import { CreateAdminDto } from './dto/create-admin.dto';
 import { UpdateAdminDto } from './dto/update-admin.dto';
 import { PrismaService } from 'prisma/prisma.service';
@@ -20,7 +20,11 @@ export class AdminService {
 
   async findOne(id: string) {
     try {
-      return await this.prisma.admins.findFirst({ where: { id: id } })
+      const admin = await this.prisma.admins.findFirst({ where: { id: id } })
+      if (!admin) {
+        return { message: "ADMIN not found", status: HttpStatus.NOT_FOUND };
+      }
+      return admin
     } catch (e) {
       console.log(e);
       return { error: e }
@@ -29,8 +33,13 @@ export class AdminService {
 
   async update(id: string, updateUserDto: UpdateAdminDto) {
     try {
+      const admin = await this.prisma.admins.findFirst({ where: { id: id } })
+      if (!admin) {
+        return { message: "ADMIN not found", status: HttpStatus.NOT_FOUND };
+      }
+
       const newUser = await this.prisma.admins.update({ data: updateUserDto, where: { id: id } });
-      return { message: "Admin updated", newUser: newUser };
+      return { message: "ADMIN updated", newUser: newUser };
     } catch (e) {
       console.log(e);
       return { error: e }
@@ -39,8 +48,13 @@ export class AdminService {
 
   async remove(id: string) {
     try {
+      const admin = await this.prisma.admins.findFirst({ where: { id: id } })
+      if (!admin) {
+        return { message: "ADMIN not found", status: HttpStatus.NOT_FOUND };
+      }
+
       const deleteUser = await this.prisma.admins.delete({ where: { id: id } });
-      return { message: "Admin deleted", deletedUser: deleteUser };
+      return { message: "ADMIN deleted", deletedUser: deleteUser };
 
     } catch (e) {
       console.log(e);
