+++ src/driver/driver.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpStatus, Injectable } from '@nestjs/common';
 import { CreateDriverDto } from './dto/create-driver.dto';
 import { UpdateDriverDto } from './dto/update-driver.dto';
 import { PrismaService } from 'prisma/prisma.service';
@@ -20,7 +20,12 @@ export class DriverService {
 
   async findOne(id: string) {
     try {
-      return await this.prisma.drivers.findFirst({ where: { id: id } })
+      const driver = await this.prisma.drivers.findFirst({ where: { id: id } })
+      if (!driver) {
+        return { messages: `DRIVER not found`, status: HttpStatus.NOT_FOUND };
+      }
+      return driver
+
     } catch (e) {
       console.log(e);
       return { error: e }
@@ -29,6 +34,12 @@ export class DriverService {
 
   async update(id: string, updateUserDto: UpdateDriverDto) {
     try {
+
+      const driver = await this.prisma.drivers.findFirst({ where: { id: id } })
+      if (!driver) {
+        return { messages: `DRIVER not found`, status: HttpStatus.NOT_FOUND };
+      }
+
       const newUser = await this.prisma.drivers.update({ data: updateUserDto, where: { id: id } });
       return { message: "Driver updated", newUser: newUser };
     } catch (e) {
@@ -39,6 +50,12 @@ export class DriverService {
 
   async remove(id: string) {
     try {
+
+      const driver = await this.prisma.drivers.findFirst({ where: { id: id } })
+      if (!driver) {
+        return { messages: `DRIVER not found`, status: HttpStatus.NOT_FOUND };
+      }
+
       const deleteUser = await this.prisma.drivers.delete({ where: { id: id } });
       return { message: "Driver deleted", deletedUser: deleteUser };
